chore: Update npm dependencies
This commit updates the npm dependencies in the package.json and package-lock.json files. It adds the following dependencies:
- @hookform/resolvers@3.6.0
- @radix-ui/react-label@2.1.0
- react-hook-form@7.52.0
- zod@3.23.8

It also updates the following dependencies to their latest versions:
- @radix-ui/react-icons
- @radix-ui/react-slot
- class-variance-authority
- clsx
- next
- react
- react-dom
- react-icons
- tailwind-merge
- tailwindcss-animate

These updates ensure that the project is using the latest stable versions of the dependencies and includes new features and bug fixes.