chore: Update npm dependencies and add next-auth package
This commit updates the npm dependencies in the package.json and package-lock.json files. It adds the following dependencies:
- @types/bcryptjs@2.4.6
- next-auth@5.0.0-beta.19

These updates ensure that the project is using the latest stable versions of the dependencies and includes new features and bug fixes. The addition of the next-auth package sets the foundation for implementing authentication functionality in the application.